✅ Exercise 3 - Dedicated worker for QR code scanning
